disallow one option for missing vertices
When not all cells have the same number of vertices, the unneeded vertices should be masked.  The option of duplicating other vertex coordinate values is now excluded because this will make it easier to do additional calculations concerning the shapes of grid cells in future releases (beyond just calculating the mean_resolution.